+++ src/coker/analysis/__init__.py
@@ -3,15 +3,17 @@
 from .controllability import analyse_controllability
 from .identifiability import analyse_identifiability
 
+from .dae import SymbolicDAESystem, lower_dae_system
 from .model import AnalysisResult, AnalysisStatus
 from .symbolic import SymbolicSystem, UnsupportedSystemError, lower_system
 
 __all__ = (
     "AnalysisResult",
     "AnalysisStatus",
-    "SymbolicSystem",
+    "SymbolicDAESystem",
     "UnsupportedSystemError",
     "analyse_controllability",
     "analyse_identifiability",
     "lower_system",
+    "lower_dae_system",
 )